+++ Classes/Domain/Validator/TypoScriptValidator.php
@@ -3,7 +3,7 @@ namespace Portrino\PxValidation\Domain\Validator;
 /***************************************************************
  *  Copyright notice
  *
- *  (c) 2015 Andre Wuttig <wuttig@portrino.de>, portrino GmbH
+ *  (c) 2016 Andre Wuttig <wuttig@portrino.de>, portrino GmbH
  *
  *  All rights reserved
  *