+++ ext_emconf.php
@@ -1,12 +1,16 @@
 <?php
 
-########################################################################
-# Extension Manager/Repository config file for ext: "px_validation"
-#
-#
-########################################################################
+/***************************************************************
+ * Extension Manager/Repository config file for ext "px_validation".
+ *
+ * Auto generated | Identifier: 979da373ac0c440fcd7cf68e0bf1e4e6
+ *
+ * Manual updates:
+ * Only the data in the array - everything else is removed by next
+ * writing. "version" and "dependencies" must not be touched!
+ ***************************************************************/
 
-$EM_CONF[$_EXTKEY] = array(
+$EM_CONF[$_EXTKEY] = array (
 	'title' => 'Extbase Validation via Typoscript',
 	'description' => 'Provides Extbase Validation via TypoScript.',
 	'category' => 'fe',
@@ -14,19 +18,25 @@ $EM_CONF[$_EXTKEY] = array(
 	'author_email' => 'wuttig@portrino.de',
 	'author_company' => 'portrino GmbH',
 	'state' => 'beta',
-	'uploadfolder' => '0',
+	'uploadfolder' => false,
 	'createDirs' => '',
 	'clearCacheOnLoad' => 0,
-	'version' => '1.0.0',
-	'constraints' => array(
-		'depends' => array(
+	'version' => '1.0.1',
+	'constraints' => 
+	array (
+		'depends' => 
+		array (
 			'typo3' => '6.2.0-7.99.99',
 		),
-		'conflicts' => array(
+		'conflicts' => 
+		array (
 		),
-		'suggests' => array(
+		'suggests' => 
+		array (
 		),
 	),
+	'clearcacheonload' => false,
+	'_md5_values_when_last_written' => 'a:30:{s:12:"CHANGELOG.md";s:4:"4957";s:7:"LICENSE";s:4:"a23a";s:9:"README.md";s:4:"6d16";s:13:"composer.json";s:4:"6322";s:12:"ext_icon.gif";s:4:"2e43";s:14:"ext_tables.php";s:4:"5a17";s:8:"Classes/";s:4:"d41d";s:15:"Classes/Domain/";s:4:"d41d";s:25:"Classes/Domain/Validator/";s:4:"d41d";s:46:"Classes/Domain/Validator/AbstractValidator.php";s:4:"c2c0";s:48:"Classes/Domain/Validator/TypoScriptValidator.php";s:4:"e4bf";s:19:"Classes/Validation/";s:4:"d41d";s:40:"Classes/Validation/ValidatorResolver.php";s:4:"4d87";s:29:"Classes/Validation/Validator/";s:4:"d41d";s:53:"Classes/Validation/Validator/ConjunctionValidator.php";s:4:"f816";s:14:"Configuration/";s:4:"d41d";s:25:"Configuration/TypoScript/";s:4:"d41d";s:34:"Configuration/TypoScript/setup.txt";s:4:"f249";s:14:"Documentation/";s:4:"d41d";s:26:"Documentation/Includes.txt";s:4:"bdfc";s:23:"Documentation/Index.rst";s:4:"c9ba";s:26:"Documentation/Settings.yml";s:4:"c91e";s:32:"Documentation/ForAdministrators/";s:4:"d41d";s:41:"Documentation/ForAdministrators/Index.rst";s:4:"b1a6";s:43:"Documentation/ForAdministrators/TypoScript/";s:4:"d41d";s:52:"Documentation/ForAdministrators/TypoScript/Index.rst";s:4:"5ff4";s:27:"Documentation/Introduction/";s:4:"d41d";s:36:"Documentation/Introduction/Index.rst";s:4:"4903";s:20:"Documentation/Todos/";s:4:"d41d";s:29:"Documentation/Todos/Index.rst";s:4:"b1ad";}',
 );
 
 ?>