+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # PxValidation Change log
 
+1.0.1 - 2016-03-22
+------------------
+* extends `composer.json` with some information
+* anonymous function to initialize the extension
+
+
 1.0.0 - 2016-01-19
 ------------------
-
 * initial release for TER